# backend/server/openapi.json
{
  "components": {
    "schemas": {
...
      "ApplyRulesResponse": {
        "properties": {
          "applied": {
            "description": "\u0414\u043e\u0431\u0430\u0432\u043b\u0435\u043d\u043e \u0438\u043b\u0438 \u0438\u0437\u043c\u0435\u043d\u0435\u043d\u043e \u043f\u0440\u0430\u0432\u0438\u043b",
            "title": "Applied",
            "type": "integer"
          },
          "errors": {
            "description": "\u041e\u0448\u0438\u0431\u043a\u0438 \u043f\u0440\u0438\u043c\u0435\u043d\u0435\u043d\u0438\u044f",
            "items": {
              "type": "string"
            },
            "title": "Errors",
            "type": "array"
          },
          "skipped": {
            "description": "\u041f\u0440\u0430\u0432\u0438\u043b \u0431\u0435\u0437 \u0438\u0437\u043c\u0435\u043d\u0435\u043d\u0438\u0439",
            "title": "Skipped",
            "type": "integer"
          }
        },
        "required": [
          "applied",
          "skipped"
        ],
        "title": "ApplyRulesResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RoutingScope": {
        "enum": [
          "host",
          "container"
        ],
        "title": "RoutingScope",
        "type": "string"
      },
      "RuleCreate": {
        "properties": {
          "country": {
            "description": "\u041a\u043e\u0434 \u0441\u0442\u0440\u0430\u043d\u044b ISO 3166-1 alpha-2",
            "title": "Country",
            "type": "string"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "fwmark": {
            "description": "\u041c\u0435\u0442\u043a\u0430 \u043f\u0430\u043a\u0435\u0442\u0430",
            "title": "Fwmark",
            "type": "integer"
          },
          "ipset_name": {
            "description": "\u0418\u043c\u044f ipset \u043d\u0430 \u0430\u0433\u0435\u043d\u0442\u0435",
            "title": "Ipset Name",
            "type": "string"
          },
          "scope": {
            "$ref": "#/components/schemas/RoutingScope",
            "default": "host",
            "description": "\u0413\u0434\u0435 \u043f\u0440\u0438\u043c\u0435\u043d\u044f\u0442\u044c (host|container)"
          },
          "scope_target": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "\u0418\u043c\u044f docker-\u043a\u043e\u043d\u0442\u0435\u0439\u043d\u0435\u0440\u0430 \u0434\u043b\u044f scope=container",
            "title": "Scope Target"
          },
          "table_id": {
            "description": "\u0422\u0430\u0431\u043b\u0438\u0446\u0430 \u043c\u0430\u0440\u0448\u0440\u0443\u0442\u0438\u0437\u0430\u0446\u0438\u0438",
            "title": "Table Id",
            "type": "integer"
          },
          "via_gateway": {
            "description": "\u0428\u043b\u044e\u0437 \u0432\u043d\u0443\u0442\u0440\u0438 \u0442\u0443\u043d\u043d\u0435\u043b\u044f",
            "title": "Via Gateway",
            "type": "string"
          },
          "via_interface": {
            "description": "\u0418\u0441\u0445\u043e\u0434\u044f\u0449\u0438\u0439 \u0438\u043d\u0442\u0435\u0440\u0444\u0435\u0439\u0441 \u043d\u0430 \u0430\u0433\u0435\u043d\u0442\u0435",
            "title": "Via Interface",
            "type": "string"
          }
        },
        "required": [
          "country",
          "ipset_name",
          "fwmark",
          "table_id",
          "via_interface",
          "via_gateway"
        ],
        "title": "RuleCreate",
        "type": "object"
      },
      "RuleListResponse": {
        "properties": {
          "rules": {
            "items": {
              "$ref": "#/components/schemas/RuleResponse"
            },
            "title": "Rules",
            "type": "array"
          }
        },
        "required": [
          "rules"
        ],
        "title": "RuleListResponse",
        "type": "object"
      },
      "RuleResponse": {
        "properties": {
          "country": {
            "title": "Country",
            "type": "string"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "fwmark": {
            "title": "Fwmark",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "ipset_name": {
            "title": "Ipset Name",
            "type": "string"
          },
          "scope": {
            "title": "Scope",
            "type": "string"
          },
          "scope_target": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope Target"
          },
          "server_id": {
            "title": "Server Id",
            "type": "integer"
          },
          "table_id": {
            "title": "Table Id",
            "type": "integer"
          },
          "via_gateway": {
            "title": "Via Gateway",
            "type": "string"
          },
          "via_interface": {
            "title": "Via Interface",
            "type": "string"
          }
        },
        "required": [
          "id",
          "server_id",
          "country",
          "ipset_name",
          "fwmark",
          "table_id",
          "via_interface",
          "via_gateway",
          "enabled",
          "scope",
          "scope_target"
        ],
        "title": "RuleResponse",
        "type": "object"
      },
      "RuleUpdate": {
        "properties": {
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          },
          "fwmark": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fwmark"
          },
          "scope": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RoutingScope"
              },
              {
                "type": "null"
              }
            ]
          },
          "scope_target": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope Target"
          },
          "table_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Table Id"
          },
          "via_gateway": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Via Gateway"
          },
          "via_interface": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Via Interface"
          }
        },
        "title": "RuleUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Waygate Control Plane",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/servers/{server_id}/rules": {
      "get": {
        "operationId": "list_rules_api_v1_servers__server_id__rules_get",
        "parameters": [
          {
            "in": "path",
            "name": "server_id",
            "required": true,
            "schema": {
              "title": "Server Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "access_token",
            "required": false,
            "schema": {
              "default": "",
              "title": "Access Token",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Rules",
        "tags": [
          "rules"
        ]
      },
      "post": {
        "operationId": "create_rule_api_v1_servers__server_id__rules_post",
        "parameters": [
          {
            "in": "path",
            "name": "server_id",
            "required": true,
            "schema": {
              "title": "Server Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "access_token",
            "required": false,
            "schema": {
              "default": "",
              "title": "Access Token",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Rule",
        "tags": [
          "rules"
        ]
      }
    },
    "/api/v1/servers/{server_id}/rules/apply": {
      "post": {
        "description": "\u0411\u0435\u0440\u0451\u0442 \u0432\u0441\u0451 \u0441\u043e\u0434\u0435\u0440\u0436\u0438\u043c\u043e\u0435 \u0442\u0430\u0431\u043b\u0438\u0446\u044b routing_rule \u0434\u043b\u044f server_id \u0438 \u0448\u043b\u0451\u0442 \u043d\u0430 \u0430\u0433\u0435\u043d\u0442.",
        "operationId": "apply_rules_api_v1_servers__server_id__rules_apply_post",
        "parameters": [
          {
            "in": "path",
            "name": "server_id",
            "required": true,
            "schema": {
              "title": "Server Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "access_token",
            "required": false,
            "schema": {
              "default": "",
              "title": "Access Token",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplyRulesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Apply Rules",
        "tags": [
          "rules"
        ]
      }
    },
    "/api/v1/servers/{server_id}/rules/{rule_id}": {
      "delete": {
        "operationId": "delete_rule_api_v1_servers__server_id__rules__rule_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "server_id",
            "required": true,
            "schema": {
              "title": "Server Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "title": "Rule Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "access_token",
            "required": false,
            "schema": {
              "default": "",
              "title": "Access Token",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Rule",
        "tags": [
          "rules"
        ]
      },
      "patch": {
        "operationId": "update_rule_api_v1_servers__server_id__rules__rule_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "server_id",
            "required": true,
            "schema": {
              "title": "Server Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "title": "Rule Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "access_token",
            "required": false,
            "schema": {
              "default": "",
              "title": "Access Token",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Rule",
        "tags": [
          "rules"
        ]
      }
    },
...
  }
}